replace original ServiceTagsAndCanaryTest with revised
